+++ test/sql/xyz_myPackage.sql
@@ -93,7 +93,7 @@ as
     -- complete
     MyPackage2.
     -- complete
-    MyPackage2.myV
+    return MyPackage2.myV
 
     return;
   end;