+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.3.4
 * Fix issue with parser and nested case...end [#47](https://github.com/zabel-xyz/plsql-language/issues/47)
+* Fix issue with the detection of a bad word during auto-completion [#48](https://github.com/zabel-xyz/plsql-language/issues/48)
 
 ## 1.3.3
 * Fix issues completion is case sensitive [#32](https://github.com/zabel-xyz/plsql-language/issues/32)